Feature partially added, plus fixed left over code.
Feature being implemented :  As a user I want to be able to search a film by name of film, name of an actor, or genre of the film
Currently doing the name of the film feature aspect so 1/3 of it.
Also removed old nonmodular code from film with actors function.