Feature Completed!
Feature Used: As a user I want to be able to search a film by name of film, name of an actor, or genre of the film

Added Package.json to the commit!
Also fixed the endspoints a bit more to align with what I want and completed the 2 sections I needed!