Final feature implemented In backend.
Feature:As a user I want to be able to rent a film out to a customer.

This feature was definitely the hardest of the 7, simply because I had to account for more than one action and had to change the structure after I finished initially since the nested structure was ugly and not readable. Ended up using async,await, promises and error handling, I probably won't go back and change the other models but for the future I'll most likely implement this structure since I like it more for how clean it looks. Now time to implement and connect to the frontend, and perform further testing after.(I've been testing if these work with MYSQL workbench and Postman)