Feature 10/14 Add customer
Feature: As a user I want to be able to add a new customer
This feature was interesting to implement because it seems simple at first glance but when you actually try to implement, you get caught by a lot of other things.

Overall in terms of the backend though, it was pretty straightforward after checking the data model/table structure.