Feature 12/14 Delete Customer
Feature: As a user I want to be able to delete a customer if they no longer wish to patron at store

Overall, the backend was pretty easy to set up once again. Simply copying the structure of how to add a customer and essentially "reverse" it and delete was up and working.